test cases
checking results with current test cases.

The problem with these values is that I did not make them with an expected tip value per hour and assigned random tips per worker. I need to design a case with the tips made per hour and calculate those values, then test and see how accurate this program is outside of the current test case