hard coded test case
In order to find how accurate the program is, I hard-coded the store tips per hour then inserted the expected tips for each worker based on the store tip income divided evenly by the number of workers. I left an error with one worker which shows a signivficant difference once the difference from the expected values are calculated.