Buckets and pigeonhole
The current proble with estimating the tips values is how tips are aggregated for each worker. if a worker starts a shift and then more workers start later in this workers shift, the sum of tips accumulated at the time are added to their total. But that total isnt calculated until their shift ends, which is when they pull the tips again. this means that the sum of the first pull 'x' and second pull when they leave 'y' equal the total that they receive, but the rate cannot be calculated without knowing what 'x' was originally. the rates a which tips earned between 'x' and 'y' are different, and cannot be solved.