Checking tip accuracy
added a bucket to track tips for current workers when a new worker joins or when someone leaves and splits tips. Currently checks when the nubmer of workers in a shift changes, but dosen't account for when someone leaves and a new worker joins in the same hour. needs revision