Bucket compare
added a bucket checking system to show what a worker should recieve based on the tips per hour, by the hour, divided by the number of workers when a new bucket is made. This is another way to check the accuracy of tips, if given the ability to count tips per hour.